Move services into `charging-module` folder
