+++ app/services/charging-module/charging-module-create-bill-run.service.js
@@ -6,10 +6,10 @@
  */
 
 const ChargingModuleTokenService = require('./charging-module-token.service.js')
-const RegionModel = require('../models/water/region.model.js')
-const RequestLib = require('../lib/request.lib.js')
+const RegionModel = require('../../models/water/region.model.js')
+const RequestLib = require('../../lib/request.lib.js')
 
-const servicesConfig = require('../../config/services.config.js')
+const servicesConfig = require('../../../config/services.config.js')
 
 /**
  * Sends a request to the Charging Module to create a new bill run and returns the result.
