+++ app/services/charging-module/charging-module-token.service.js
@@ -5,9 +5,9 @@
  * @module ChargingModuleTokenService
  */
 
-const RequestLib = require('../lib/request.lib.js')
+const RequestLib = require('../../lib/request.lib.js')
 
-const servicesConfig = require('../../config/services.config.js')
+const servicesConfig = require('../../../config/services.config')
 
 async function go () {
   const url = new URL('/oauth2/token', servicesConfig.chargingModule.token.url)
